Add tests for PUT, POST, DELETE
These tests are not 100% stable yet as they now make changes to the
server data that affect other tests. To counter this, we also now
recommend using the "--watch" flag with json-server, so that you can at
least manually revert the files and re-run the tests without having to
restart the server.

Subsequent commits will improve the tests so that this will no longer be
necessary.