Fix test for removing an element
This is a separate commit to show what that JavaScript is difficult to
write even with a good IDE. The querySelector() method claims that it
returns a "map", but extracting the first element with array
deconstruction didn't work as expected. The call to querySelectorAll()
returned an iterator, but the call to removeChild() didn't complain
about it.

With better typing, this could be avoided, of course. We use these tests
to verify that our code works as expected.