Fix test-interdependency in event test suite
The test for stopPropagation() left an event listener attached to the
document that captured all events and stopped propagation. The test for
preventDefault() would no longer run correctly with that listener
attached. The solution is to make sure that the capturing document
listener is removed at the end of the stopPropagation() test.