Generalize all components to use metadata
The MasterDetail, List, and Detail components no longer make assumptions
about the type of data they use. Instead, the components reference a
simple set of metadata to determine which URL to use to access data,
which properties the type has, how to show a title for entities, and how
to initialize a new object.