React: Use controlled inputs in the detail
The defaultValue should only be used with uncontrolled components.
Because our inputs now have an onChange handler, they are considered
controlled. On top of that, we should be updating the model for each
property change in order to best integrate with React's state-handling.

See the following link for a longer explanation:

https://scriptverse.academy/tutorials/reactjs-defaultvalue-value.html