Move list-item and input-row components top-level
We do this because if only a single list-item or input-row needs to be
re-rendered, the entire parent component was re-rendered with it. If the
render functions are not nested, then React doesn't have to render the
parent component in order to render the child.

See the following link for more information:

https://dev.to/adevnadia/how-to-write-performant-react-code-rules-patterns-dos-and-donts-40gk

It summarizes this as Rule #3.

"Never create new components inside the render function of another component."