Fix incorrect usage of async in useEffect()
Although the function appeared to be working just fine, the type-checker
says that the signature doesn't match. This is correct. It is known that
useEffect() doesn't handle async, so you have to use the workaround in
this commit to get it work properly.

See the following links for more information.

- https://devtrium.com/posts/async-functions-useeffect
- https://www.benmvp.com/blog/successfully-using-async-functions-useeffect-react/