Use "as const" and ReturnType<> to define AppAction
Instead of managing interfaces by hand, we define functions to create
the desired actions for the reducer. We then use the ReturnType<> to
get the type of the function result and take a discriminated union of
those instead. This yields a fully typed action with the added bonus
that we can use the functions to create the actions from the service.
On top of that, we can also add data to an action simply by adding a
parameter and/or including more data in the function result.